should run now

+++ parcels/lead-java/src/main/java/lead/freightDemand/FreightFacility.java
@@ -20,7 +20,7 @@ public class FreightFacility {
     private final Coord coord;
     private int st8 = 0;
     private String st45 = "";
-    double movements = 0;
+    int movements = 0;
     double centerDistance = 0;
 
     public FreightFacility(String siret, String activitePrincipaleEtablissement, String trancheEffectifsEtablissement, String siren, Coord coord) throws Exception {
